+++ R/distributions.R
@@ -4653,7 +4653,7 @@ tfd_sample_distribution <- function(distribution,
 #' @param dtype_override samples of distributions will be cast to this dtype. If
 #'  unspecified, all distributions must have the same dtype. Default value:
 #'  `NULL` (i.e., do not cast).
-#'
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
 #' @export
 tfd_blockwise <- function(distributions,
                           dtype_override=NULL,
@@ -4686,7 +4686,7 @@ tfd_blockwise <- function(distributions,
 #'  The absolute tolerance for comparing closeness to loc. Default is 0.
 #' @param rtol Non-negative Tensor of same dtype as loc and broadcastable shape.
 #'  The relative tolerance for comparing closeness to loc. Default is 0.
-#'
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
 #' @export
 tfd_vector_deterministic <- function(loc,
                                      atol = NULL,
@@ -4720,7 +4720,7 @@ tfd_vector_deterministic <- function(loc,
 #'  batch of independent distributions and the last dimension represents a vector
 #'  of probabilities for each class. Only one of logits or probs should be passed
 #'  in.
-#'
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
 #' @export
 tfd_exp_relaxed_one_hot_categorical <- function(temperature,
                                                 logits = NULL,
@@ -4785,6 +4785,7 @@ tfd_exp_relaxed_one_hot_categorical <- function(temperature,
 #' @param scale Floating point tensor; the scales of the distribution.
 #' Must contain only positive values.
 #' @param name string prefixed to Ops created by this class. Default value: 'doublesided_maxwell'.
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
 #' @export
 tfd_doublesided_maxwell <- function(loc,
                                     scale,
@@ -4835,6 +4836,7 @@ tfd_doublesided_maxwell <- function(loc,
 #' floating number comparison. Only effective when `outcomes` is a floating
 #' `Tensor`. Default is `10 * eps`.
 #' @param name string prefixed to Ops created by this class.
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
 #' @export
 tfd_finite_discrete <- function(outcomes,
                                 logits = NULL,
@@ -4924,6 +4926,7 @@ tfd_finite_discrete <- function(outcomes,
 #' magnitude, the more the distribution concentrates near `loc` (for
 #' `concentration >= 0`) or near `loc - (scale/concentration)` (for
 #' `concentration < 0`). Floating point `Tensor`.
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
 #' @export
 tfd_generalized_pareto <- function(loc,
                                    scale,
@@ -4950,6 +4953,7 @@ tfd_generalized_pareto <- function(loc,
 #' Normal distribution.
 #'
 #' @inherit tfd_normal return params
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
 #' @export
 tfd_logit_normal <- function(loc,
                              scale,
@@ -4971,7 +4975,7 @@ tfd_logit_normal <- function(loc,
 #' whose logarithm is normally distributed with mean `loc` and
 #' standard deviation `scale`. It is constructed as the exponential
 #' transformation of a Normal distribution.
-
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
 #' @inherit tfd_normal return params
 #' @export
 tfd_log_normal <- function(loc,
@@ -5023,6 +5027,7 @@ tfd_log_normal <- function(loc,
 #' @param high upper bound
 #' @param temperature controls the shape of the distribution
 #' @inherit tfd_normal return params
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
 #' @export
 tfd_pert <- function(low,
                      peak,
@@ -5100,6 +5105,7 @@ tfd_pert <- function(low,
 #' vector of scores for the elements.
 #' @param dtype The type of the event samples (default: int32).
 #' @inherit tfd_normal return params
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
 #' @export
 tfd_plackett_luce <- function(scores,
                               dtype = tf$int32,
@@ -5142,6 +5148,7 @@ tfd_plackett_luce <- function(scores,
 #' @param dtype The type of the event samples. Default: `int32`.
 #'
 #' @family distributions
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
 #' @export
 tfd_probit_bernoulli <- function(probits = NULL,
                                  probs = NULL,
@@ -5161,4 +5168,216 @@ tfd_probit_bernoulli <- function(probits = NULL,
   do.call(tfp$distributions$ProbitBernoulli, args)
 }
 
-## https://github.com/tensorflow/probability/blob/master/tensorflow_probability/python/distributions/pixel_cnn.py
+#' The matrix Wishart distribution on positive definite matrices
+#'
+#' This distribution is defined by a scalar number of degrees of freedom df and
+#' an instance of LinearOperator, which provides matrix-free access to a
+#' symmetric positive definite operator, which defines the scale matrix.
+#'
+#' Mathematical Details
+#'
+#' The probability density function (pdf) is,
+#' ```
+#' pdf(X; df, scale) = det(X)**(0.5 (df-k-1)) exp(-0.5 tr[inv(scale) X]) / Z
+#' Z = 2**(0.5 df k) |det(scale)|**(0.5 df) Gamma_k(0.5 df)
+#' ```
+#' where:
+#' * `df >= k` denotes the degrees of freedom,
+#' * `scale` is a symmetric, positive definite, `k x k` matrix,
+#' * `Z` is the normalizing constant, and,
+#' * `Gamma_k` is the [multivariate Gamma function](
+#'  https://en.wikipedia.org/wiki/Multivariate_gamma_function).
+#'
+#' @param df float or double tensor, the degrees of freedom of the
+#' distribution(s). df must be greater than or equal to k.
+#' @param scale `float` or `double` instance of `LinearOperator`.
+#' @param input_output_cholesky Logical. If TRUE, functions whose input or
+#' output have the semantics of samples assume inputs are in Cholesky form
+#' and return outputs in Cholesky form. In particular, if this flag is
+#' TRUE, input to log_prob is presumed of Cholesky form and output from
+#' sample, mean, and mode are of Cholesky form.  Setting this
+#' argument to TRUE is purely a computational optimization and does not
+#' change the underlying distribution; for instance, mean returns the
+#' Cholesky of the mean, not the mean of Cholesky factors. The variance
+#' and stddev methods are unaffected by this flag.
+#' Default value: FALSE (i.e., input/output does not have Cholesky semantics).
+#'
+#' @inherit tfd_normal return params
+#'
+#' @family distributions
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
+#' @export
+tfd_wishart_linear_operator <- function(df,
+                                        scale,
+                                        input_output_cholesky = FALSE,
+                                        validate_args = FALSE,
+                                        allow_nan_stats = TRUE,
+                                        name = "WishartLinearOperator") {
+  args <- list(
+    df = df,
+    scale = scale,
+    input_output_cholesky = input_output_cholesky,
+    validate_args = validate_args,
+    allow_nan_stats = allow_nan_stats,
+    name = name
+  )
+
+  do.call(tfp$distributions$WishartLinearOperator, args)
+}
+
+#' The matrix Wishart distribution parameterized with Cholesky factors.
+#'
+#' This distribution is defined by a scalar degrees of freedom `df` and a scale
+#' matrix, expressed as a lower triangular Cholesky factor.
+#'
+#' Mathematical Details
+#'
+#' The probability density function (pdf) is,
+#' ```
+#' pdf(X; df, scale) = det(X)**(0.5 (df-k-1)) exp(-0.5 tr[inv(scale) X]) / Z
+#' Z = 2**(0.5 df k) |det(scale)|**(0.5 df) Gamma_k(0.5 df)
+#' ```
+#' where:
+#' * `df >= k` denotes the degrees of freedom,
+#' * `scale` is a symmetric, positive definite, `k x k` matrix,
+#' * `Z` is the normalizing constant, and,
+#' * `Gamma_k` is the [multivariate Gamma function](
+#'  https://en.wikipedia.org/wiki/Multivariate_gamma_function).
+#'
+#' @param df float or double tensor, the degrees of freedom of the
+#' distribution(s). df must be greater than or equal to k.
+#' @param scale_tril `float` or `double` `Tensor`. The Cholesky factorization
+#' of the symmetric positive definite scale matrix of the distribution.
+#' @param input_output_cholesky Logical. If TRUE, functions whose input or
+#' output have the semantics of samples assume inputs are in Cholesky form
+#' and return outputs in Cholesky form. In particular, if this flag is
+#' TRUE, input to log_prob is presumed of Cholesky form and output from
+#' sample, mean, and mode are of Cholesky form.  Setting this
+#' argument to TRUE is purely a computational optimization and does not
+#' change the underlying distribution; for instance, mean returns the
+#' Cholesky of the mean, not the mean of Cholesky factors. The variance
+#' and stddev methods are unaffected by this flag.
+#' Default value: FALSE (i.e., input/output does not have Cholesky semantics).
+#'
+#' @inherit tfd_normal return params
+#'
+#' @family distributions
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
+#' @export
+tfd_wishart_tri_l <- function(df,
+                              scale_tril,
+                              input_output_cholesky = FALSE,
+                              validate_args = FALSE,
+                              allow_nan_stats = TRUE,
+                              name = "WishartTriL") {
+  args <- list(
+    df = df,
+    scale_tril = scale_tril,
+    input_output_cholesky = input_output_cholesky,
+    validate_args = validate_args,
+    allow_nan_stats = allow_nan_stats,
+    name = name
+  )
+
+  do.call(tfp$distributions$WishartTriL, args)
+}
+
+#' The Pixel CNN++ distribution
+#'
+#' Pixel CNN++ (Salimans et al., 2017) models a distribution over image
+#' data, parameterized by a neural network. It builds on Pixel CNN and
+#' Conditional Pixel CNN, as originally proposed by
+#' (van den Oord et al., 2016).
+#' The model expresses the joint distribution over pixels as
+#' the product of conditional distributions:
+#' `p(x|h) = prod{ p(x[i] | x[0:i], h) : i=0, ..., d }`, in which
+#' `p(x[i] | x[0:i], h) : i=0, ..., d` is the
+#' probability of the `i`-th pixel conditional on the pixels that preceded it in
+#' raster order (color channels in RGB order, then left to right, then top to
+#' bottom). `h` is optional additional data on which to condition the image
+#' distribution, such as class labels or VAE embeddings. The Pixel CNN++
+#' network enforces the dependency structure among pixels by applying a mask to
+#' the kernels of the convolutional layers that ensures that the values for each
+#' pixel depend only on other pixels up and to the left.
+#' Pixel values are modeled with a mixture of quantized logistic distributions,
+#' which can take on a set of distinct integer values (e.g. between 0 and 255
+#' for an 8-bit image).
+#' Color intensity `v` of each pixel is modeled as:
+#' `v ~ sum{q[i] * quantized_logistic(loc[i], scale[i]) : i = 0, ..., k }`,
+#' in which `k` is the number of mixture components and the `q[i]` are the
+#' Categorical probabilities over the components.
+#'
+#' @section References:
+#' - [Tim Salimans, Andrej Karpathy, Xi Chen, and Diederik P. Kingma. PixelCNN++: Improving the PixelCNN with Discretized Logistic Mixture Likelihood and Other Modifications. In _International Conference on Learning Representations_, 2017.](https://pdfs.semanticscholar.org/9e90/6792f67cbdda7b7777b69284a81044857656.pdf)
+#' - [Aaron van den Oord, Nal Kalchbrenner, Oriol Vinyals, Lasse Espeholt, Alex Graves, and Koray Kavukcuoglu. Conditional Image Generation with PixelCNN Decoders. In _Neural Information Processing Systems_, 2016.](https://arxiv.org/abs/1606.05328)
+#' - [Aaron van den Oord, Nal Kalchbrenner, and Koray Kavukcuoglu. Pixel Recurrent Neural Networks. In _International Conference on Machine Learning_, 2016.](https://arxiv.org/pdf/1601.06759.pdf)
+#'
+#' @param image_shape 3D `TensorShape` or tuple for the `[height, width, channels]`
+#' dimensions of the image.
+#' @param conditional_shape `TensorShape` or tuple for the shape of the
+#' conditional input, or `NULL` if there is no conditional input.
+#' @param num_resnet `integer`, the number of layers (shown in Figure 2 of https://arxiv.org/abs/1606.05328)
+#' within each highest-level block of Figure 2 of https://pdfs.semanticscholar.org/9e90/6792f67cbdda7b7777b69284a81044857656.pdf.
+#' @param num_hierarchies `integer`, the number of hightest-level blocks (separated by
+#' expansions/contractions of dimensions in Figure 2 of https://pdfs.semanticscholar.org/9e90/6792f67cbdda7b7777b69284a81044857656.pdf.)
+#' @param num_filters `integer`, the number of convolutional filters.
+#' @param num_logistic_mix `integer`, number of components in the logistic mixture
+#' distribution.
+#' @param receptive_field_dims `tuple`, height and width in pixels of the receptive
+#' field of the convolutional layers above and to the left of a given
+#' pixel. The width (second element of the tuple) should be odd. Figure 1
+#' (middle) of https://arxiv.org/abs/1606.05328 shows a receptive field of (3, 5)
+#' (the row containing the current pixel is included in the height).
+#' The default of (3, 3) was used to produce the results in https://pdfs.semanticscholar.org/9e90/6792f67cbdda7b7777b69284a81044857656.pdf.
+#' @param dropout_p `float`, the dropout probability. Should be between 0 and 1.
+#' @param resnet_activation `string`, the type of activation to use in the resnet blocks.
+#' May be 'concat_elu', 'elu', or 'relu'.
+#' @param use_weight_norm `logical`, if `TRUE` then use weight normalization (works
+#' only in Eager mode).
+#' @param use_data_init `logical`, if `TRUE` then use data-dependent initialization
+#' (has no effect if `use_weight_norm` is `FALSE`).
+#' @param high `integer`, the maximum value of the input data (255 for an 8-bit image).
+#' @param low `integer`, the minimum value of the input data.
+#' @param dtype Data type of the `Distribution`.
+#' @param name `string`, the name of the `Distribution`.
+#' @inherit tfd_normal return params
+#'
+#' @family distributions
+#' @seealso For usage examples see e.g. [tfd_sample()], [tfd_log_prob()], [tfd_mean()].
+#' @export
+tfd_pixel_cnn <- function(image_shape,
+                          conditional_shape = NULL,
+                          num_resnet = 5,
+                          num_hierarchies = 3,
+                          num_filters = 160,
+                          num_logistic_mix = 10,
+                          receptive_field_dims = c(3, 3),
+                          dropout_p = 0.5,
+                          resnet_activation = 'concat_elu',
+                          use_weight_norm = TRUE,
+                          use_data_init = TRUE,
+                          high = 255,
+                          low = 0,
+                          dtype = tf$float32,
+                          name = 'PixelCNN') {
+  args <- list(
+    image_shape = normalize_shape(image_shape),
+    conditional_shape = normalize_shape(conditional_shape),
+    num_resnet = as.integer(num_resnet),
+    num_hierarchies = as.integer(num_hierarchies),
+    num_filters = as.integer(num_filters),
+    num_logistic_mix = as.integer(num_logistic_mix),
+    receptive_field_dims = normalize_shape(receptive_field_dims),
+    dropout_p = dropout_p,
+    resnet_activation = resnet_activation,
+    use_weight_norm = use_weight_norm,
+    use_data_init = use_data_init,
+    high = as.integer(high),
+    low = as.integer(low),
+    dtype = dtype,
+    name = name
+  )
+
+  do.call(tfp$distributions$PixelCNN, args)
+}
+