+++ R/sts.R
@@ -1353,3 +1353,203 @@ sts_constrained_seasonal_state_space_model <-
     )
     do.call(tfp$sts$ConstrainedSeasonalStateSpaceModel, args)
   }
+
+#' Formal representation of a smooth seasonal effect model
+#'
+#' The smooth seasonal model uses a set of trigonometric terms in order to
+#' capture a recurring pattern whereby adjacent (in time) effects are
+#' similar. The model uses `frequencies` calculated via:
+#'
+#' ```
+#' frequencies[j] = 2. * pi * frequency_multipliers[j] / period
+#' ```
+#' and then posits two latent states for each `frequency`. The two latent states
+#' associated with frequency `j` drift over time via:
+#'
+#' ```
+#' effect[t] = (effect[t-1] * cos(frequencies[j]) +
+#'              auxiliary[t-] * sin(frequencies[j]) +
+#'              Normal(0., drift_scale))
+#' auxiliary[t] = (-effect[t-1] * sin(frequencies[j]) +
+#'                 auxiliary[t-] * cos(frequencies[j]) +
+#'                 Normal(0., drift_scale))
+#' ```
+#'
+#' where `effect` is the smooth seasonal effect and `auxiliary` only appears as a
+#' matter of construction. The interpretation of `auxiliary` is thus not
+#' particularly important.
+#' @param period positive scalar `float` `Tensor` giving the number of timesteps
+#' required for the longest cyclic effect to repeat.
+#' @param frequency_multipliers One-dimensional `float` `Tensor` listing the
+#' frequencies (cyclic components) included in the model, as multipliers of
+#' the base/fundamental frequency `2. * pi / period`. Each component is
+#' specified by the number of times it repeats per period, and adds two
+#' latent dimensions to the model. A smooth seasonal model that can
+#' represent any periodic function is given by
+#' `frequency_multipliers = [1,2, ..., floor(period / 2)]`.
+#' However, it is often desirable to enforce a
+#' smoothness assumption (and reduce the computational burden) by dropping
+#' some of the higher frequencies.
+#' @param allow_drift optional `logical` specifying whether the seasonal
+#' effects can drift over time.  Setting this to `FALSE`
+#' removes the `drift_scale` parameter from the model. This is
+#' mathematically equivalent to
+#' `drift_scale_prior = tfd.Deterministic(0.)`, but removing drift
+#' directly is preferred because it avoids the use of a degenerate prior.
+#' Default value: `TRUE`.
+#' @param drift_scale_prior optional `tfd$Distribution` instance specifying a prior
+#' on the `drift_scale` parameter. If `NULL`, a heuristic default prior is
+#' constructed based on the provided `observed_time_series`.
+#' Default value: `NULL`.
+#' @param initial_state_prior instance of `tfd$MultivariateNormal` representing
+#' the prior distribution on the latent states. Must have event shape
+#' `[2 * len(frequency_multipliers)]`. If `NULL`, a heuristic default prior
+#' is constructed based on the provided `observed_time_series`.
+#' @param observed_time_series optional `float` `Tensor` of shape
+#' `batch_shape + [T, 1]` (omitting the trailing unit dimension is also
+#' supported when `T > 1`), specifying an observed time series.
+#' Any priors not explicitly set will be given default values according to
+#' the scale of the observed time series (or batch of time series). May
+#' optionally be an instance of `tfp$sts$MaskedTimeSeries`, which includes
+#' a mask `Tensor` to specify timesteps with missing observations.
+#' Default value: `NULL`.
+#'
+#' @inherit sts_local_linear_trend return params
+#' @family sts
+#' @seealso For usage examples see [sts_fit_with_hmc()], [sts_forecast()], [sts_decompose_by_component()].
+#'
+#' @export
+sts_smooth_seasonal <- function(period,
+                                frequency_multipliers,
+                                allow_drift = TRUE,
+                                drift_scale_prior = NULL,
+                                initial_state_prior = NULL,
+                                observed_time_series = NULL,
+                                name = NULL) {
+  args <- list(
+    period = period,
+    frequency_multipliers = frequency_multipliers,
+    allow_drift = allow_drift,
+    drift_scale_prior = drift_scale_prior,
+    initial_state_prior = initial_state_prior,
+    observed_time_series = observed_time_series,
+    name = name
+  )
+
+  do.call(tfp$sts$SmoothSeasonal, args)
+
+}
+
+#' State space model for a smooth seasonal effect
+#'
+#' A state space model (SSM) posits a set of latent (unobserved) variables that
+#' evolve over time with dynamics specified by a probabilistic transition model
+#' `p(z[t+1] | z[t])`. At each timestep, we observe a value sampled from an
+#' observation model conditioned on the current state, `p(x[t] | z[t])`. The
+#' special case where both the transition and observation models are Gaussians
+#' with mean specified as a linear function of the inputs, is known as a linear
+#' Gaussian state space model and supports tractable exact probabilistic
+#' calculations; see `tfp$distributions$LinearGaussianStateSpaceModel` for
+#' details.
+#' A smooth seasonal effect model is a special case of a linear Gaussian SSM. It
+#' is the sum of a set of "cyclic" components, with one component for each
+#' frequency:
+#' ```
+#' frequencies[j] = 2. * pi * frequency_multipliers[j] / period
+#' ```
+#' Each cyclic component contains two latent states which we denote `effect` and
+#' `auxiliary`. The two latent states for component `j` drift over time via:
+#' ```
+#' effect[t] = (effect[t-1] * cos(frequencies[j]) +
+#'              auxiliary[t-] * sin(frequencies[j]) +
+#'              Normal(0., drift_scale))
+#' auxiliary[t] = (-effect[t-1] * sin(frequencies[j]) +
+#'                 auxiliary[t-] * cos(frequencies[j]) +
+#'                 Normal(0., drift_scale))
+#' ```
+#'
+#' The `auxiliary` latent state only appears as a matter of construction and thus
+#' its interpretation is not particularly important. The total smooth seasonal
+#' effect is the sum of the `effect` values from each of the cyclic components.
+#' The parameters `drift_scale` and `observation_noise_scale` are each (a batch
+#' of) scalars. The batch shape of this `Distribution` is the broadcast batch
+#' shape of these parameters and of the `initial_state_prior`.
+#'
+#' Mathematical Details
+#'
+#' The smooth seasonal effect model implements a
+#' `tfp$distributions$LinearGaussianStateSpaceModel` with
+#' `latent_size = 2 * len(frequency_multipliers)` and `observation_size = 1`.
+#' The latent state is the concatenation of the cyclic latent states which themselves
+#' comprise an `effect` and an `auxiliary` state. The transition matrix is a block diagonal
+#' matrix where block `j` is:
+#' ```
+#' transition_matrix[j] =  [[cos(frequencies[j]), sin(frequencies[j])],
+#'                          [-sin(frequencies[j]), cos(frequencies[j])]]
+#' ```
+#'
+#' The observation model picks out the cyclic `effect` values from the latent state:
+#' ```
+#' observation_matrix = [[1., 0., 1., 0., ..., 1., 0.]]
+#' observation_noise ~ Normal(loc=0, scale=observation_noise_scale)
+#' ```
+#' For further mathematical details please see Harvey (1990).
+#'
+#' @section references:
+#' - Harvey, A. Forecasting, Structural Time Series Models and the Kalman Filter. Cambridge: Cambridge University Press, 1990.
+#'
+#' @param num_timesteps Scalar `integer` `Tensor` number of timesteps to model
+#' with this distribution.
+#' @param period positive scalar `float` `Tensor` giving the number of timesteps
+#' required for the longest cyclic effect to repeat.
+#' @param frequency_multipliers One-dimensional `float` `Tensor` listing the
+#' frequencies (cyclic components) included in the model, as multipliers of
+#' the base/fundamental frequency `2. * pi / period`. Each component is
+#' specified by the number of times it repeats per period, and adds two
+#' latent dimensions to the model. A smooth seasonal model that can
+#' represent any periodic function is given by
+#' `frequency_multipliers = [1,2, ..., floor(period / 2)]`.
+#' However, it is often desirable to enforce a
+#' smoothness assumption (and reduce the computational burden) by dropping
+#' some of the higher frequencies.
+#' @param drift_scale Scalar (any additional dimensions are treated as batch
+#' dimensions) `float` `Tensor` indicating the standard deviation of the
+#' latent state transitions.
+#' @param initial_state_prior instance of `tfd$MultivariateNormal`
+#' representing the prior distribution on latent states.  Must have
+#' event shape `[num_features]`.
+#' @param observation_noise_scale Scalar (any additional dimensions are
+#' treated as batch dimensions) `float` `Tensor` indicating the standard
+#' deviation of the observation noise. Default value: `0.`.
+#' @param initial_step scalar `integer` `Tensor` specifying the starting timestep.
+#' Default value: `0`.
+#'
+#' @inherit sts_local_linear_trend_state_space_model return params
+#' @family sts
+#'
+#' @export
+sts_smooth_seasonal_state_space_model <-
+  function(num_timesteps,
+           period,
+           frequency_multipliers,
+           drift_scale,
+           initial_state_prior,
+           observation_noise_scale = 0,
+           initial_step = 0,
+           validate_args = FALSE,
+           allow_nan_stats = TRUE,
+           name = NULL) {
+    args <- list(
+      num_timesteps = as.integer(num_timesteps),
+      period = period,
+      frequency_multipliers = frequency_multipliers,
+      drift_scale = drift_scale,
+      initial_state_prior = initial_state_prior,
+      observation_noise_scale = observation_noise_scale,
+      initial_step = as.integer(initial_step),
+      validate_args = validate_args,
+      allow_nan_stats = allow_nan_stats,
+      name = name
+    )
+    do.call(tfp$sts$SmoothSeasonalStateSpaceModel, args)
+  }