+++ NEWS.md
@@ -10,6 +10,9 @@
   - logit_normal
   - log_normal
   - pert
+  - tfd_wishart_linear_operator
+  - tfd_wishart_tri_l
+  - tfd_pixel_cnn
   
 - new bijectors
   - tfb_shift
@@ -25,6 +28,10 @@
   - tfb_kumaraswamy_cdf
   - tfb_fill_scale_triangular
   - tfb_ffjord
+  
+- new state space models
+  - sts_smooth_seasonal
+  - sts_smooth_seasonal_state_space_model
 
 
 ## 0.8.0.0 (Initial release, CRAN)